+++ src/ULDP/ukrr.py
@@ -4,7 +4,7 @@ from collections import Counter
 
 
 class UKRR(object):
-    def __init__(self, epsilon: int, domain: list, data: list, xs: list, xn: list):
+    def __init__(self, epsilon: float, domain: list, data: list, xs: list, xn: list):
         super(UKRR, self).__init__()
         # 隐私预算
         self.epsilon = epsilon
